google_sheets_api/requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
dateparser==1.1.7
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
oauthlib==3.2.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tzdata==2022.7
tzlocal==4.2
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.14
